Delete Unnecessary Parts
