+++ src/assets/main.css
@@ -9,7 +9,6 @@
 
 body {
   margin: 0;
-  /*overflow: hidden;*/
 }
 
 a {
