+++ src/views/TestView.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang="ts">
-import {ref,reactive} from 'vue';
+import {ref} from 'vue';
 import { useRoute } from 'vue-router';
 const route = useRoute();
 const word = ref(route.params.word);
